reverted some changes to Graphics, made the drawable shapes use double instead of int for most of their variables

+++ include/Wrapper/Graphics/Shapes/CircleShape.hpp
@@ -6,8 +6,8 @@ class GDIWRAPPER_API CircleShape : public EllipseShape
 public:
 	CircleShape();
 	CircleShape(double radius);
-	CircleShape(int x, int y, double radius);
-	CircleShape(std::pair<int, int> pos, double radius);
+	CircleShape(double x, double y, double radius);
+	CircleShape(std::pair<double, double> pos, double radius);
 
 	void setRadius(double radius);
